Moved push_folder_to_github method to the utils.py file from oecd_data.py and iati_rhfp.py files
