+++ data_updates/Python/oecd_data.py
@@ -5,7 +5,7 @@ import glob
 import numpy
 from numpy import NaN
 import pandas as pd
-from github import Github, InputGitTreeElement
+
 
 current_date = datetime.datetime.now()
 PURPOSE_CODE_TRENDS_URL = os.getenv('PURPOSE_CODE_TRENDS_URL', "https://staging-ddw.devinit.org/api/export/1241/")
@@ -17,36 +17,7 @@ REMOTE_FOLDER = f'{current_date.year}'
 ODA_AID_TYPE_URL = os.getenv('ODA_AID_TYPE_URL', "https://staging-ddw.devinit.org/api/export/1238/")
 ODA_CHANNEL_TYPE_URL = os.getenv('ODA_CHANNEL_TYPE_URL', "https://staging-ddw.devinit.org/api/export/1237/")
 ODA_RECIP_TYPE_URL = os.getenv('ODA_RECIP_TYPE_URL', "https://ddw.devinit.org/api/export/1339/")
-GITHUB_TOKEN = os.getenv('GITHUB_TOKEN', None)
 
-def push_folder_to_github(repo_name, branch, local_folder, remote_folder, commit_msg, file_extension='*.*'):
-    dir_path = os.path.dirname(os.path.realpath(__file__))
-    glob_path = os.path.join(dir_path, local_folder, file_extension)
-    csv_paths = glob.glob(glob_path)
-    abs_csv_paths = [os.path.abspath(csv) for csv in csv_paths]
-    if len(abs_csv_paths) < 1:
-        print('Nothing to push')
-        return
-    g = Github(GITHUB_TOKEN)
-    # Uncomment below line to use your token for testing. Make sure you do not push changes with your token to Github
-    # g = Github(GITHUB_KEY)
-    repo = g.get_repo(repo_name)
-    git_ref = repo.get_git_ref('heads/' + branch)
-    branch_obj = repo.get_branch(branch=branch)
-    git_sha = branch_obj.commit.sha
-    base_tree = repo.get_git_tree(git_sha)
-    element_list = list()
-    for abs_csv_path in abs_csv_paths:
-        with open(abs_csv_path) as input_file:
-            data = input_file.read()
-        remote_file = remote_folder + '/' + os.path.basename(abs_csv_path)
-        element = InputGitTreeElement(remote_file, '100644', 'blob', data)
-        element_list.append(element)
-
-    tree = repo.create_git_tree(element_list, base_tree)
-    parent = repo.get_git_commit(git_sha)
-    commit = repo.create_git_commit(commit_msg, tree, [parent])
-    git_ref.edit(commit.sha)
 
 # create csv folder if it does not exist
 path = 'data_updates/Python/oecd_csv'
